Added list and segment search filter
Added search box that filters the visible list of trail segments.

In order to enable filtering of map lines based on the segment search feature, code was refactored such that Google map elements are now part of the Segment data model (rather than the controller/ViewModel). By default, map segment lines are set to show, and when search filters are applied, the filtered segments are set to hide.