Added time period change functionality
Now, users can select a time period of interest for trail traffic. The default option is "current" and includes a count of any Strava-recorded cyclists in the past few hours. The other options give average traffic for days of the week (average daily traffic over the last 3 instances of a given weekday).

Still to-do: add color scales to map lines that indicate severity of traffic.